requirements.txt
streamlit
google-generativeai
SpeechRecognition
gTTS
Pillow
pyaudio